docs: server functions use one config, not a separate provider list
Removes the `serverFnAppProviders` convention from the docs, RFC, and
demo now that handlers bootstrap from the app's server config:

- the guide registers interceptors and `withAllowedOrigins` in
  `app.config.server.ts`, and states that `providedIn: 'root'` services
  and app-config providers resolve with no extra registration
- the RFC's provider-source and DI-parity sections describe the
  one-config bootstrap; test-coverage notes the app-injector spec
- the demo moves its interceptor to `app.config.server.ts` and deletes
  `server-fns/index.ts`; the harnesses build their injector from the
  interceptor the same way
