+++ apps/analog-app/scripts/validate-server-fn.ts
@@ -13,16 +13,21 @@ import '@angular/platform-server/init';
 import {
   dispatchServerFn,
   createServerFnAppInjector,
+  provideServerFns,
+  withServerFnInterceptors,
 } from '@analogjs/router/server';
-import { serverFnAppProviders } from '../src/app/server-fns';
+import { authInterceptor } from '../src/app/server-fns/auth.interceptor';
 import { ids, registerServerFns } from './_server-fn-harness';
 
 // The app injector is built ONCE (as the generated Nitro handler does). Each
 // dispatch passes only `{ parent }` — no per-request providers — so the handler
 // resolving CatalogService + interceptors proves they come from the parent.
-// Bootstrapped app injector, exactly as the generated Nitro handler builds it,
-// so a `providedIn: 'root'` service (CatalogService) resolves without listing.
-const appInjector = await createServerFnAppInjector(serverFnAppProviders);
+// Bootstrapped app injector, as the generated Nitro handler builds it from the
+// app's server config: a `providedIn: 'root'` service (CatalogService) resolves
+// without listing, and the interceptor comes from the config's providers.
+const appInjector = await createServerFnAppInjector([
+  ...provideServerFns(withServerFnInterceptors([authInterceptor])),
+]);
 
 function fakeEvent(headers: Record<string, string> = {}) {
   return { node: { req: { headers }, res: {} } } as any;
